Enable the docker image to use memcached
The terrain server running within the container can be configured to use a
memcache server to cache tileset data by either specifying a memcached container
to link to or setting an environment variable.

Any container running memcached on port 11211 and linked with the alias
`memcached` will be used.  E.g. assume the following memcached container is
running:

    docker run --name memcache -d memcached

This can then be used by a terrain server image:

    docker run --name terrain -d --link memcache:memcached geodata/cesium-terrain-server

A memcached server that is not linked can be used by setting the container
`MEMCACHED` environment variable to point to the memcached network address e.g.

    docker run --name terrain -d --env MEMCACHED=memcache.me.org:11211 geodata/cesium-terrain-server