Remove the reverse proxy
The terrain server can now serve up all the static Cesium assets as well as the
tilesets so it makes sense to decouple the container from the reverse proxy and
simplify things.  The terrain server can populate memcache caches, which an
external reverse proxy would be able to make use of.